Rename api_reference to types
